alfred/Alfred.alfredpreferences/workflows/yelp/requirements-minimal.txt
addict
Alfred-Workflow
beautifulsoup4
requests
